+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2023-05-04-1
+
+Changes:
+
+- Constants
+- Input events
+
 ## v2023-05-01-1
 
 Changes: